[MO] - [refactor] -> utils
