+++ systemtest/src/main/java/io/strimzi/systemtest/logs/TestExecutionWatcher.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.utils;
+package io.strimzi.systemtest.logs;
 
 import io.strimzi.systemtest.Environment;
 import org.apache.logging.log4j.LogManager;
