+++ systemtest/src/test/java/io/strimzi/systemtest/AbstractST.java
@@ -39,7 +39,7 @@ import io.strimzi.api.kafka.model.KafkaUser;
 import io.strimzi.api.kafka.model.status.Condition;
 import io.strimzi.systemtest.clients.lib.KafkaClient;
 import io.strimzi.systemtest.interfaces.TestSeparator;
-import io.strimzi.systemtest.utils.TestExecutionWatcher;
+import io.strimzi.systemtest.logs.TestExecutionWatcher;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.HelmClient;
 import io.strimzi.test.k8s.exceptions.KubeClusterException;
@@ -581,7 +581,7 @@ public abstract class AbstractST extends BaseST implements TestSeparator {
 
     /**
      * Wait till all pods in specific namespace being deleted and recreate testing environment in case of some pods cannot be deleted.
-     * This method is {@Deprecated} and it should be removed in the future. Instead of use this method, you should use method {@link io.strimzi.systemtest.utils.StUtils#waitForPodDeletion(String)}
+     * This method is {@Deprecated} and it should be removed in the future. Instead of use this method, you should use method {@link io.strimzi.systemtest.utils.kubeUtils.objects.PodUtils#waitForPodDeletion(String)}
      * which should be used by default in deletion process for all components (it force pod deletion via executor instead of fabric8 client, which seems to be unstable).
      * @param time timeout in miliseconds
      * @throws Exception exception
